fixed: can't catch new following/followerList data problem
