+++ src/views/UserFollowList.vue
@@ -19,8 +19,6 @@
         <FollowerNavPills />
         <!-- UserFollowers.vue, UserFollowings.vue -->
         <router-view
-          :followerList="followerList"
-          :followingList="followingList"
           class="bottom-lists scrollbar"
         />
       </div>
@@ -49,8 +47,6 @@ export default {
   data() {
     return {
       topPopular: [],
-      followerList: [],
-      followingList: [],
       tweets: [],
       userName: ''  // 渲染頁面上方標題
     };
@@ -58,55 +54,9 @@ export default {
   created() {
     const { id: userId } = this.$route.params;
     this.fetchPopular();
-    this.fetchFollowers(userId);
-    this.fetchFollowings(userId)
     this.fetchTweets(userId)
   },
   methods: {
-    async fetchFollowers(userId) {
-      try {
-        const response = await userAPI.getFollowers({ userId })
-        const { data } = response
-        if (response.statusText !== 'OK') {
-          throw new Error (data.message)
-        }
-        this.followerList = data
-      } catch (error) {
-        console.error(error.message)
-        // 沒有跟隨者的情況
-        if (error.message === 'No followers found.') {
-          this.followingList = []
-          return
-        } else {
-          Toast.fire({
-            icon: 'error',
-            title: '無法取得跟隨者資料'
-          })
-        }
-      }
-    },
-    async fetchFollowings(userId) {
-      try {
-        const response = await userAPI.getFollowings( {userId} )
-        const { data } = response
-        if (response.statusText !== 'OK') {
-          throw new Error (data.message)
-        }
-        this.followingList = data
-      } catch(error) {
-        console.error(error.message)
-        // 沒有跟隨中對象的情況
-        if (error.message === 'No followings found.') {
-          this.followingList = []
-          return
-        } else {
-          Toast.fire({
-            icon: 'error',
-            title: '無法取得跟隨中的使用者資料'
-          })
-        }
-      }
-    },
     // 利用使用者 id 取得所有推文，計算推文數量使用
     async fetchTweets(userId) {
       try {
@@ -117,7 +67,6 @@ export default {
           throw new Error (data.message)
         }
         this.tweets = data;
-        console.log(data)
         this.userName = data[0].User.name
       } catch (error) {
         console.error(error.message)
